fix: rename curio to sp
