+++ packages/synapse-core/src/index.ts
@@ -12,11 +12,11 @@
 export * as abis from './abis/index.ts'
 export * as chains from './chains.ts'
 export * from './constants.ts'
-export * as curio from './curio.ts'
 export * as erc20 from './erc20.ts'
 export * as errors from './errors/index.ts'
 export * as pay from './pay/index.ts'
 export * as sessionKey from './session-key/index.ts'
+export * as curio from './sp.ts'
 export * as typedData from './typed-data/index.ts'
 export * as usdfc from './usdfc.ts'
 export * from './utils/decode-pdp-errors.ts'
